Change IFieldDefinition.Offset type from uint to int

+++ src/Compilers/Core/Portable/PEWriter/MetadataWriter.cs
@@ -3097,7 +3097,7 @@ namespace Microsoft.Cci
 
                 uint fieldDefIndex = (uint)this.GetFieldDefIndex(fieldDef);
                 FieldLayoutRow r = new FieldLayoutRow();
-                r.Offset = fieldDef.Offset;
+                r.Offset = (uint)fieldDef.Offset;
                 r.Field = fieldDefIndex;
                 _fieldLayoutTable.Add(r);
             }
